refactor: moved unprocessed_variable_update_function where it belongs
